Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%f %f %c",&a,&b,&c);
     switch(c){
-        case '+':printf("%d",a+b);break;
-        case '-':printf("%d",a-b);break;
-        case '*':printf("%d",a*b);break;
-        case '/':printf("%d",a/b);break;
+        case '+':printf("%f",a+b);break;
+        case '-':printf("%f",a-b);break;
+        case '*':printf("%f",a*b);break;
+        case '/':printf("%f",a/b);break;
         default:printf("error");
     }
